Glance V2: restruct image resource and data source code
