+++ openstack/resource_openstack_images_image_v2.go
@@ -1,23 +1,18 @@
 package openstack
 
 import (
-	"crypto/md5"
-	"encoding/hex"
 	"fmt"
-	"io"
 	"log"
-	"net/http"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/gophercloud/gophercloud"
 	"github.com/gophercloud/gophercloud/openstack/imageservice/v2/imagedata"
 	"github.com/gophercloud/gophercloud/openstack/imageservice/v2/images"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform/helper/validation"
 )
 
 func resourceImagesImageV2() *schema.Resource {
@@ -45,17 +40,21 @@ func resourceImagesImageV2() *schema.Resource {
 			},
 
 			"container_format": {
-				Type:         schema.TypeString,
-				Required:     true,
-				ForceNew:     true,
-				ValidateFunc: resourceImagesImageV2ValidateContainerFormat,
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+				ValidateFunc: validation.StringInSlice([]string{
+					"ami", "ari", "aki", "bare", "ovf", "ova",
+				}, false),
 			},
 
 			"disk_format": {
-				Type:         schema.TypeString,
-				Required:     true,
-				ForceNew:     true,
-				ValidateFunc: resourceImagesImageV2ValidateDiskFormat,
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+				ValidateFunc: validation.StringInSlice([]string{
+					"ami", "ari", "aki", "vhd", "vmdk", "raw", "qcow2", "vdi", "iso",
+				}, false),
 			},
 
 			"file": {
@@ -87,7 +86,7 @@ func resourceImagesImageV2() *schema.Resource {
 				Type:         schema.TypeInt,
 				Optional:     true,
 				ForceNew:     true,
-				ValidateFunc: validatePositiveInt,
+				ValidateFunc: validation.IntAtLeast(0),
 				Default:      0,
 			},
 
@@ -95,7 +94,7 @@ func resourceImagesImageV2() *schema.Resource {
 				Type:         schema.TypeInt,
 				Optional:     true,
 				ForceNew:     true,
-				ValidateFunc: validatePositiveInt,
+				ValidateFunc: validation.IntAtLeast(0),
 				Default:      0,
 			},
 
@@ -127,11 +126,13 @@ func resourceImagesImageV2() *schema.Resource {
 			},
 
 			"visibility": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ForceNew:     false,
-				ValidateFunc: resourceImagesImageV2ValidateVisibility,
-				Default:      "private",
+				Type:     schema.TypeString,
+				Optional: true,
+				ForceNew: false,
+				ValidateFunc: validation.StringInSlice([]string{
+					"public", "private", "shared", "community",
+				}, false),
+				Default: "private",
 			},
 
 			"properties": {
@@ -305,8 +306,6 @@ func resourceImagesImageV2Read(d *schema.ResourceData, meta interface{}) error {
 	d.Set("metadata", img.Metadata)
 	d.Set("created_at", img.CreatedAt.Format(time.RFC3339))
 	d.Set("updated_at", img.UpdatedAt.Format(time.RFC3339))
-	// Deprecated
-	d.Set("update_at", img.UpdatedAt.Format(time.RFC3339))
 	d.Set("container_format", img.ContainerFormat)
 	d.Set("disk_format", img.DiskFormat)
 	d.Set("min_disk_gb", img.MinDiskGigabytes)
@@ -318,6 +317,9 @@ func resourceImagesImageV2Read(d *schema.ResourceData, meta interface{}) error {
 	d.Set("tags", img.Tags)
 	d.Set("visibility", img.Visibility)
 	d.Set("region", GetRegion(d, config))
+
+	// Deprecated
+	d.Set("update_at", img.UpdatedAt.Format(time.RFC3339))
 
 	properties := resourceImagesImageV2ExpandProperties(img.Properties)
 	if err := d.Set("properties", properties); err != nil {
@@ -444,234 +446,5 @@ func resourceImagesImageV2Delete(d *schema.ResourceData, meta interface{}) error
 	}
 
 	d.SetId("")
-	return nil
-}
-
-func resourceImagesImageV2ValidateVisibility(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	validVisibilities := []string{
-		"public",
-		"private",
-		"shared",
-		"community",
-	}
-
-	for _, v := range validVisibilities {
-		if value == v {
-			return
-		}
-	}
-
-	err := fmt.Errorf("%s must be one of %s", k, validVisibilities)
-	errors = append(errors, err)
-	return
-}
-
-func validatePositiveInt(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(int)
-	if value >= 0 {
-		return
-	}
-	errors = append(errors, fmt.Errorf("%q must be a positive integer", k))
-	return
-}
-
-var DiskFormats = [9]string{"ami", "ari", "aki", "vhd", "vmdk", "raw", "qcow2", "vdi", "iso"}
-
-func resourceImagesImageV2ValidateDiskFormat(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	for i := range DiskFormats {
-		if value == DiskFormats[i] {
-			return
-		}
-	}
-	errors = append(errors, fmt.Errorf("%q must be one of %v", k, DiskFormats))
-	return
-}
-
-var ContainerFormats = [9]string{"ami", "ari", "aki", "bare", "ovf", "ova"}
-
-func resourceImagesImageV2ValidateContainerFormat(v interface{}, k string) (ws []string, errors []error) {
-	value := v.(string)
-	for i := range ContainerFormats {
-		if value == ContainerFormats[i] {
-			return
-		}
-	}
-	errors = append(errors, fmt.Errorf("%q must be one of %v", k, ContainerFormats))
-	return
-}
-
-func resourceImagesImageV2MemberStatusFromString(v string) images.ImageMemberStatus {
-	switch v {
-	case "accepted":
-		return images.ImageMemberStatusAccepted
-	case "pending":
-		return images.ImageMemberStatusPending
-	case "rejected":
-		return images.ImageMemberStatusRejected
-	case "all":
-		return images.ImageMemberStatusAll
-	}
-
-	return ""
-}
-
-func resourceImagesImageV2VisibilityFromString(v string) images.ImageVisibility {
-	switch v {
-	case "public":
-		return images.ImageVisibilityPublic
-	case "private":
-		return images.ImageVisibilityPrivate
-	case "shared":
-		return images.ImageVisibilityShared
-	case "community":
-		return images.ImageVisibilityCommunity
-	}
-
-	return ""
-}
-
-func fileMD5Checksum(f *os.File) (string, error) {
-	hash := md5.New()
-	if _, err := io.Copy(hash, f); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(hash.Sum(nil)), nil
-}
-
-func resourceImagesImageV2FileProps(filename string) (int64, string, error) {
-	var filesize int64
-	var filechecksum string
-
-	file, err := os.Open(filename)
-	if err != nil {
-		return -1, "", fmt.Errorf("Error opening file for Image: %s", err)
-
-	}
-	defer file.Close()
-
-	fstat, err := file.Stat()
-	if err != nil {
-		return -1, "", fmt.Errorf("Error reading image file %q: %s", file.Name(), err)
-	}
-
-	filesize = fstat.Size()
-	filechecksum, err = fileMD5Checksum(file)
-
-	if err != nil {
-		return -1, "", fmt.Errorf("Error computing image file %q checksum: %s", file.Name(), err)
-	}
-
-	return filesize, filechecksum, nil
-}
-
-func resourceImagesImageV2File(d *schema.ResourceData) (string, error) {
-	if filename := d.Get("local_file_path").(string); filename != "" {
-		return filename, nil
-	} else if furl := d.Get("image_source_url").(string); furl != "" {
-		dir := d.Get("image_cache_path").(string)
-		os.MkdirAll(dir, 0700)
-		filename := filepath.Join(dir, fmt.Sprintf("%x.img", md5.Sum([]byte(furl))))
-
-		if _, err := os.Stat(filename); err != nil {
-			if !os.IsNotExist(err) {
-				return "", fmt.Errorf("Error while trying to access file %q: %s", filename, err)
-			}
-			log.Printf("[DEBUG] File doens't exists %s. will download from %s", filename, furl)
-			file, err := os.Create(filename)
-			if err != nil {
-				return "", fmt.Errorf("Error creating file %q: %s", filename, err)
-			}
-			defer file.Close()
-			resp, err := http.Get(furl)
-			if err != nil {
-				return "", fmt.Errorf("Error downloading image from %q", furl)
-			}
-			defer resp.Body.Close()
-
-			if _, err = io.Copy(file, resp.Body); err != nil {
-				return "", fmt.Errorf("Error downloading image %q to file %q: %s", furl, filename, err)
-			}
-			return filename, nil
-		} else {
-			log.Printf("[DEBUG] File exists %s", filename)
-			return filename, nil
-		}
-	} else {
-		return "", fmt.Errorf("Error in config. no file specified")
-	}
-}
-
-func resourceImagesImageV2RefreshFunc(client *gophercloud.ServiceClient, id string) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		img, err := images.Get(client, id).Extract()
-		if err != nil {
-			return nil, "", err
-		}
-		log.Printf("[DEBUG] OpenStack image status is: %s", img.Status)
-
-		return img, fmt.Sprintf("%s", img.Status), nil
-	}
-}
-
-func resourceImagesImageV2BuildTags(v []interface{}) []string {
-	var tags []string
-	for _, tag := range v {
-		tags = append(tags, tag.(string))
-	}
-
-	return tags
-}
-
-func resourceImagesImageV2ExpandProperties(v map[string]interface{}) map[string]string {
-	properties := map[string]string{}
-	for key, value := range v {
-		if v, ok := value.(string); ok {
-			properties[key] = v
-		}
-	}
-
-	return properties
-}
-
-func resourceImagesImageV2UpdateComputedAttributes(diff *schema.ResourceDiff, meta interface{}) error {
-	if diff.HasChange("properties") {
-		// Only check if the image has been created.
-		if diff.Id() != "" {
-			// Try to reconcile the properties set by the server
-			// with the properties set by the user.
-			//
-			// old = user properties + server properties
-			// new = user properties only
-			o, n := diff.GetChange("properties")
-
-			newProperties := resourceImagesImageV2ExpandProperties(n.(map[string]interface{}))
-
-			for oldKey, oldValue := range o.(map[string]interface{}) {
-				// os_ keys are provided by the OpenStack Image service.
-				if strings.HasPrefix(oldKey, "os_") {
-					if v, ok := oldValue.(string); ok {
-						newProperties[oldKey] = v
-					}
-				}
-
-				// direct_url is provided by some storage drivers.
-				if oldKey == "direct_url" {
-					if v, ok := oldValue.(string); ok {
-						newProperties[oldKey] = v
-					}
-				}
-			}
-
-			// Set the diff to the newProperties, which includes the server-side
-			// os_ properties.
-			//
-			// If the user has changed properties, they will be caught at this
-			// point, too.
-			diff.SetNew("properties", newProperties)
-		}
-	}
-
 	return nil
 }
